+++ examples/text/text_raylib_fonts.c
@@ -52,8 +52,8 @@ int main(void)
 
     for (int i = 0; i < MAX_FONTS; i++)
     {
-        positions[i].x = screenWidth/2.0f - MeasureTextEx(fonts[i], messages[i], fonts[i].baseSize*2.0f, spacings[i]).x/2.0f;
-        positions[i].y = 60 + fonts[i].baseSize + 45*i;
+        positions[i].x = screenWidth/2.0f - MeasureTextEx(fonts[i], messages[i], fonts[i].baseSize*2.0f, (float)spacings[i]).x/2.0f;
+        positions[i].y = 60.0f + fonts[i].baseSize + 45.0f*i;
     }
 
     // Small Y position corrections
